Tweak WindowConf types.

+++ crates/nvi/src/types.rs
@@ -560,9 +560,9 @@ pub struct WindowConf {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bufpos: Option<(u64, u64)>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub row: Option<f64>,
+    pub row: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub col: Option<f64>,
+    pub col: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub focusable: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
